snapshot: combine nogc snapshotting helpers
